==> processing_scripts/5p/wt/01-ambient_RNA_check.runfile.Rmd <==
---
title: 'scRNA QC'
author: "FHIL\nDerrikGratz"
date: '`r Sys.Date()`'
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      fig.width = 8, fig.height = 6,
                      cache=FALSE, cache.lazy = FALSE) 

library(tidyverse)  ## General R logic
library(here)       ## Easier specification of file locations
library(yaml)       ## parses config yaml
library(Seurat)
library(SoupX)
library(reshape2)   ## DF manipulation
library(data.table) ## rbindlist
library(immunarch)
# library(FHIL.SC.utils) ## Custom functions
set.seed(33)
```

```{r}
figures <- list()
```

# Functions

```{r}
runSoupX <- function(tod,
                     toc,
                     clusters,
                     doPlot = TRUE,
                     roundToInt = TRUE){
  ## this is intended for scale data. They have outputs with additional barcodes somehow
  if (!all(colnames(toc) %in% names(clusters)) | !(all(names(clusters) %in% colnames(toc)))){
    warning('Names of clusters and filtered data to not match. Subsetting to values present in both')
    toc_full <- toc
    toc <- toc[,colnames(toc)[colnames(toc) %in% names(clusters)]]
    clusters <- clusters[names(clusters)[names(clusters) %in% colnames(toc)]]
    store_full = TRUE
  } else {
    store_full = FALSE
  }
  sc <- SoupX::SoupChannel(tod = tod,
                           toc = toc[,names(clusters)])
  sc <- SoupX::setClusters(sc, clusters)
  sc <- SoupX::autoEstCont(sc, doPlot = doPlot, contaminationRange = c(0.001,0.8))
  sc$adjusted_counts <- SoupX::adjustCounts(sc, roundToInt = roundToInt)
  sc$plot <- grDevices::recordPlot()
  
  if (store_full) {
    sc$toc_full <- toc_full
  }
  sc
}
```

# Load config

```{r}
source(here('config/5p/generate_metadata.R'))
metadata <- read.csv(here('config/5p/metadata.csv'))
metadata
```

```{r}
objs <- list()
```

# Load WT data

## gemx_5p

```{r}
kit <- 'gemx_5p'
for (i in 1:nrow(samplesheets[[kit]])){
  sample <- samplesheets[[kit]][i,'Sample']
  file <- samplesheets[[kit]][i,'File']
  
  filtered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file,
                               'outs', 'per_sample_outs', file, 'count/sample_filtered_feature_bc_matrix.h5')
  filtered_counts <- Read10X_h5(filtered_counts)
  
  unfiltered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file,
                                 'outs', 'multi/count/raw_feature_bc_matrix.h5')
  unfiltered_counts <- Read10X_h5(unfiltered_counts)
  
  clusters <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'outs', 'per_sample_outs',
                        file, 'count/analysis/clustering/gene_expression_graphclust/clusters.csv')
  clusters <- read.csv(clusters)
  clusters <- setNames(as.list(clusters$Cluster), clusters$Barcode)
  objs[[sample]] <- list(tod = unfiltered_counts, toc = filtered_counts, clusters = clusters)
}
```

## nextgem_5p

```{r}
kit <- 'nextgem_5p'
for (i in 1:nrow(samplesheets[[kit]])){
  sample <- samplesheets[[kit]][i,'Sample']
  file <- samplesheets[[kit]][i,'File']
  
  filtered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file,
                               'outs', 'per_sample_outs', file, 'count/sample_filtered_feature_bc_matrix.h5')
  filtered_counts <- Read10X_h5(filtered_counts)
  
  unfiltered_counts <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file,
                                 'outs', 'multi/count/raw_feature_bc_matrix.h5')
  unfiltered_counts <- Read10X_h5(unfiltered_counts)
  
  clusters <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], file, 'outs', 'per_sample_outs',
                        file, 'count/analysis/clustering/gene_expression_graphclust/clusters.csv')
  clusters <- read.csv(clusters)
  clusters <- setNames(as.list(clusters$Cluster), clusters$Barcode)
  objs[[sample]] <- list(tod = unfiltered_counts, toc = filtered_counts, clusters = clusters)
}
```

## parse_v2

```{r}
kit <- 'parse_v2'
for (i in 1:nrow(samplesheets[[kit]])){
  sample <- samplesheets[[kit]][i,'Sample']
  file <- samplesheets[[kit]][i,'File']
  
  filtered_counts <- file.path(config$base_dir, config$data_dir,  data_dirs[[kit]],'WT/Combined/', file, 'DGE_filtered')
  filtered_counts <- ReadParseBio(filtered_counts)
  
  unfiltered_counts <- file.path(config$base_dir, config$data_dir,  data_dirs[[kit]], 'WT/Combined/',file, 'DGE_unfiltered')
  unfiltered_counts <- ReadParseBio(unfiltered_counts)
  
  clusters <- file.path(config$base_dir, config$data_dir, data_dirs[[kit]], 'WT/Combined/', file, 'report/cluster_assignment.csv')
  clusters <- read.csv(clusters)
  clusters <- setNames(as.list(clusters$cluster), clusters$bc_wells)
  
  objs[[sample]] <- list(tod = unfiltered_counts, toc = filtered_counts, clusters = clusters)
}
```

# SoupX

```{r}
objs <- lapply(objs, function(x) {
  runSoupX(x$tod, x$toc, x$clusters, doPlot = TRUE, roundToInt = TRUE)
})
```

## Save RDS

```{r, eval=FALSE}
saveRDS(objs, here('rds/5p/wt/01-soup_channels.rds'), compress=FALSE)
```


```{r, eval=TRUE}
objs <- readRDS(here('rds/5p/wt/01-soup_channels.rds'))
```

## Append individual reruns

```{r, eval=FALSE}
objs_rerun <- objs
objs <- readRDS(here('rds/5p/wt/01-soup_channels.rds'))
for (x in names(objs_rerun)) {
  objs[[x]] <- objs_rerun[[x]]
}
saveRDS(objs, here('rds/5p/wt/01-soup_channels.rds'), compress=FALSE)
```

## Summarize contamination

```{r}
extractSoupXContamEst <- function(sc){
  rho <- sc$fit$rhoEst
  rho_low <- sc$fit$rhoFWHM[1]
  rho_high <- sc$fit$rhoFWHM[2]
  return(list(rho_low = rho_low, rho = rho, rho_high = rho_high))
}
soupx_results <- data.table::rbindlist(lapply(objs, extractSoupXContamEst), idcol = 'Sample')
soupx_results
```

```{r}
figures[['rho_barchart']] <-
soupx_results %>%
  # reshape2::melt() %>%
  merge(metadata, by = 'Sample') %>%
  ggplot(aes(x=paste0(Individual, Replicate), y=rho)) +
  geom_col() +
  geom_errorbar(aes(ymin=rho_low, ymax=rho_high), width=0.4) +
  facet_wrap(~ Kit, scales='free_x', ncol=3) +
  labs(x='Sample', y='Contamination fraction')
figures[['rho_barchart']] 
```

## Correct counts (if desired)

```{r, eval = FALSE}
objs <- lapply(objs, function(x) x$adjusted_counts)
```

```{r, eval = TRUE}
objs <- lapply(objs, function(x) {
  if (!is.null(x$toc_full)) {
    x$toc_full
  } else {
    x$toc
}})
```

# Save objs

```{r}
objs <- lapply(objs, CreateSeuratObject)
for (obj in names(objs)){
  Project(objs[[obj]]) <- obj
}
for (obj in names(objs)){
  objs[[obj]]$orig.ident <- obj
}
```

```{r}
saveRDS(objs, here('rds/5p/wt/01_raw_objs.rds'), compress=FALSE)
```

# Render report

```{r, eval=TRUE}
## This only works if all the variables called in the format file 
## are in the current environment. Check the format file to tweak
## aesthetics 
rmarkdown::render(here('processing_scripts/5p/wt/01-ambient_RNA_check.format.Rmd'),
                  output_file = '01_soupx.html',
                  output_dir = here('reports/5p/wt'))
```
