==> processing_scripts/5p/wt/03-gene_analysis.format.Rmd <==
---
title: 'BM02 Gene set analysis'
author: |
  | FHIL
  | Derrik Gratz
date: '`r Sys.Date()`'
output: 
  rmdformats::robobook:
    fig_width: 9
    toc_depth: 3
    css: '../../../config/robobook.css'
---

```{r, include=FALSE}
library(knitr)
knitr::opts_chunk$set(echo = FALSE, warning=FALSE,
                      fig.width = 8, fig.height = 6,
                      cache=FALSE, cache.lazy = FALSE) 
```

# Filtering

Filtering for genes found in 10+ cells

```{r}
figures[['usable_genes_sample']]
```

Genes found in 10+ cells in 2+ samples for each kit. 

```{r}
figures[['usable_genes_kit']]
```

# Gene set overlap

```{r, fig.width=5, fig.height=5}
figures[['gene_overlap_euler']]
```

Same data, quantified

```{r, fig.width=16, fig.height=6}
figures[['gene_overlap_upset']]
```

# Gene expression sets

Portion of reads for each sample coming from the shared set of genes in all non-probe based kits (e.g. not Flex). 
Gene families are defined by prefix patterns of genes that were highly expressed. There may be more or more interesting
families to look at, these are just some I came up with while inspecting the top hits.

<details><summary>Gene family definitions</summary>

```{r}
gene_families
```

</details>

```{r, fig.width=10}
figures[['Expression_barchart_shared']]
```

Now looking at genes that were not part of the shared pool. These are not necessarily genes that are unique to each
kit, just those that weren't found in all 3' capture kits. 

```{r, fig.width=10}
figures[['Expression_barchart_unique']]
```

# Unique expression 

With the unique gene recovery, how much expression is being detected? While most genes look low, this is what most genes look like when quantified this way, so it's not necessarily indicating that these kits are only 'barely' picking up extra genes. Mostly just looking if 
there are any highly expressed genes that may be of interest. Here labels are provided for genes with with high expression (greater than
50% of cells or with median log-expression > 2) that don't belong to one of the pre-defined families. 

```{r, fig.width=16, fig.height=14}
figures[['unique_expression_grid']]
```
