==> processing_scripts/5p/wt/04-reference_annotation.runfile.Rmd <==
---
title: 'SC processing'
author: "FHIL\nDerrikGratz"
date: '`r Sys.Date()`'
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      fig.width = 8, fig.height = 6,
                      cache=FALSE, cache.lazy = FALSE) 

library(tidyverse)  ## General R logic
library(here)       ## Easier specification of file locations
library(yaml)       ## parses config yaml
library(Seurat)
library(reshape2)   ## DF manipulation
# library(clustree)   ## Resolution selection for clustering
# library(ggraph)     ## To render clustree
# library(bluster)     ## Sihouette scores
library(SingleR)    ## reference based annotation
library(ggforestplot) ## Striped lines for dotplots
library(SingleCellExperiment)
library(gridExtra)  ## arrange plots
library(janitor) ## Tabyl, adorn_rounding
library(ggrastr) ## Rasterize images for smaller size
library(SeuratData) ## reference datasets
rbindlist <- data.table::rbindlist
set.seed(33)
```

# Load data

```{r}
objs <- readRDS(here('rds/5p/wt/03-objs_post_gene_filtering.rds'))
metadata <- read.csv(here('config/5p/metadata.csv'))
figures <- list()
```

# Data normalization and processing

```{r}
# objs <- lapply(objs, NormalizeData)
objs <- lapply(objs, FindVariableFeatures)
objs <- lapply(objs, ScaleData)
objs <- lapply(objs, RunPCA)
```

```{r}
#Check if 20 dims is appropriate
lapply(objs, ElbowPlot, ndim=30)
```

```{r}
objs <- lapply(objs, FindNeighbors, dims=1:20, return.neighbor=TRUE)
objs <- lapply(objs, FindNeighbors, dims=1:20, return.neighbor=FALSE)
objs <- lapply(objs, RunUMAP, nn.name='RNA.nn')#dims=1:20)
```

# clustering (fixed resolution)

```{r}
objs <- lapply(objs, FindClusters, resolution=1.0, cluster.name='res_1.0_clusters') #, algorithm=4
```

```{r}
lapply(objs, DimPlot)
```

# Reference annotation

```{r run_SingleR}
runSingleR <- function(obj.seurat,
                       ref,
                       labels,
                       de.method = "classic",
                       meta.prefix = 'singleR',
                       num.threads = BiocParallel::bpnworkers(BiocParallel::MulticoreParam()),
                       ...) {
  singleR.out <- obj.seurat %>%
    Seurat::as.SingleCellExperiment(assay = "RNA") %>%
    scuttle::logNormCounts() %>%
    SummarizedExperiment::assays() # %>%
    # .$logcounts %>%
  singleR.out <- SingleR::SingleR(test = singleR.out$logcounts,
                     ref = ref,
                     labels = labels,
                     de.method = de.method,
                     num.threads = num.threads,
                     ...)

  obj.seurat@meta.data[[paste0(meta.prefix,".labels")]] <- singleR.out$labels
  obj.seurat@meta.data[[paste0(meta.prefix,".pruned.labels")]] <- singleR.out$pruned.labels
  obj.seurat@meta.data[[paste0(meta.prefix,".delta.next")]] <- singleR.out$delta.next

  return(obj.seurat)
}
```

## Novershtern

```{r}
ref.nove <- readRDS('/fh/fast/_IRC/FHIL/pub/ref/single_cell_references/Novershtern_celldex.rds')
objs <- lapply(objs, function(x) {
  runSingleR(x, 
             ref = assays(ref.nove)$logcounts,
             labels = ref.nove$label.main,
             meta.prefix = 'Novershtern.main',
             de.method = 'wilcox')
  })
```

## HPCA

```{r}
ref.hpca <- readRDS('/fh/fast/_IRC/FHIL/pub/ref/single_cell_references/HPCA_celldex.rds')
objs <- lapply(objs, function(x) {
  runSingleR(x, 
             ref = assays(ref.hpca)$logcounts,
             labels = ref.hpca$label.main,
             meta.prefix = 'HPCA.main',
             de.method = 'wilcox')
  })
```

## Monaco

```{r}
ref.mona <- readRDS('/fh/fast/_IRC/FHIL/pub/ref/single_cell_references/Monaco_celldex.rds')
objs <- lapply(objs, function(x) {
  runSingleR(x, 
             ref = assays(ref.mona)$logcounts,
             labels = ref.mona$label.main,
             meta.prefix = 'Mona.main',
             de.method = 'wilcox')
  })
```

# Seurat

## PBMC3k

```{r}
library(SeuratData)
# InstallData('pbmc3k')
library(pbmc3k.SeuratData)
data("pbmc3k")
pbmc3k <- UpdateSeuratObject(pbmc3k) %>%
  NormalizeData() %>%
  FindVariableFeatures() %>%
  ScaleData() %>%
  RunPCA()
```

```{r}
objs <- lapply(objs, function(obj) {
  anchors <- FindTransferAnchors(
    reference = pbmc3k,
    query = obj,
    normalization.method = "LogNormalize",
    reduction = "pcaproject",
    reference.reduction = 'pca',
    dims = 1:20
  )
  obj <- MapQuery(
    anchorset = anchors,
    query = obj,
    reference = pbmc3k,
    refdata = list(
      pbmc3k_seurat_annotations = "seurat_annotations"
    )#,
    # new.reduction.name = 'ref.umap'
  )
  obj
})
```

```{r,fig.width=12, fig.height=10}
# grid.arrange(DimPlot(obj, group.by = 'predicted.pbmc3k_seurat_annotations', label=TRUE),
# DimPlot(obj, group.by = 'Novershtern.main.labels', label=TRUE),
# DimPlot(obj, group.by = 'Mona.main.labels', label=TRUE),
# DimPlot(obj, group.by = 'HPCA.main.labels', label=TRUE), nrow=2)

```


## pbmcsca

```{r}
library(pbmcsca.SeuratData)
data("pbmcsca")
pbmcsca <- UpdateSeuratObject(pbmcsca) %>%
  NormalizeData() %>%
  FindVariableFeatures() %>%
  ScaleData() %>%
  RunPCA() |> 
  RunUMAP(dims=1:20, return.model = TRUE)
DimPlot(pbmcsca, group.by='CellType')
```

```{r}
objs <- lapply(objs, function(obj) {
  anchors <- FindTransferAnchors(
    reference = pbmcsca,
    query = obj,
    normalization.method = "LogNormalize",
    reduction = "pcaproject",
    reference.reduction = 'pca',
    dims = 1:20
  )
  obj <- MapQuery(
    anchorset = anchors,
    query = obj,
    reference = pbmcsca,
    refdata = list(
      pbmcsca_seurat_annotations = "CellType"
    ),
    reduction.model = "umap"
    # projectumap.args = list(
    #   # reference.reduction = 'umap',
    #   # new.reduction.name = 'pbmcsca.umap'
    # )
  )
  obj
})
```

```{r,fig.width=12, fig.height=10}
obj <- objs[[1]]
grid.arrange(DimPlot(obj, group.by = 'predicted.pbmc3k_seurat_annotations', label=TRUE),
DimPlot(obj, group.by = 'predicted.pbmcsca_seurat_annotations', label=TRUE),
DimPlot(obj, group.by = 'Mona.main.labels', label=TRUE),
DimPlot(obj, group.by = 'HPCA.main.labels', label=TRUE), nrow=2)

```

# Harmonize labels

```{r}
label_mappings <- list(
  predicted.pbmcsca_seurat_annotations = c(
    "Cytotoxic T cell" = "CD8+ T", 
    "CD4+ T cell"= "CD4+ T",
    "B cell"= "B",
    "CD14+ monocyte"= "Classical monocyte",
    "Natural killer cell"= "NK",
    "CD16+ monocyte"= "Non-classical monocyte",
    "Megakaryocyte"= "Megakaryocyte",
    "Dendritic cell"= "Dendritic",
    "Plasmacytoid dendritic cell"= "pDC"
  ),
  predicted.pbmc3k_seurat_annotations = c(
    "CD8 T"='CD8+ T',
    "Memory CD4 T" = 'CD4+ T',
    "B" = 'B',
    "CD14+ Mono" = "Classical monocyte",
    "NK" = "NK",
    "Naive CD4 T" = "CD4+ T",
    "FCGR3A+ Mono" = "Non-classical monocyte",
    "Platelet" = "Megakaryocyte",
    "DC" = "Dendritic"
  ),
  Mona.main.labels = c(
    "CD8+ T cells" = "CD8+ T",
    "CD4+ T cells" = "CD4+ T",
    "B cells" = "B",
    "Monocytes" = "Monocyte",
    "Dendritic cells" = "Dendritic",
    "T cells" = "T",
    "NK cells" = "NK",
    "Progenitors" = "Lymphocyte progenitor",
    "Basophils" = "Granulocyte",
    "Neutrophils" = "Granulocyte"
  ),
  Novershtern.main.labels = c(
    "CD8+ T cells" = "CD8+ T",
    "CD4+ T cells" = "CD4+ T",
    "Basophils" = "Granulocyte",
    "Monocytes" = "Monocyte",
    "NK cells" = "NK",
    "B cells" = "B",
    "Granulocytes" = "Granulocyte",
    "NK T cells" = "Other T",
    "Megakaryocytes" = "Megakaryocyte",
    "Dendritic cells" = "Dendritic",
    "HSCs" = "HSPC",
    "Erythroid cells" = "Erythrocyte", 
    "Eosinophils" = "Granulocyte"
  ),
  HPCA.main.labels = c(
    "T_cells" = "T",
    "B_cell" = "B",
    "Monocyte" = "Monocyte",
    "NK_cell" = "NK",
    "HSC_-G-CSF" = "HSPC",
    "Neutrophils" = "Granulocyte",
    "Platelets" = "Megakaryocyte",
    "Pre-B_cell_CD34-" = "Other B",
    "GMP" = "Myeloid",
    "CMP" = "Myeloid",
    "Myelocyte" = "Granulocyte",
    "DC" = "Dendritic",
    "Pro-B_cell_CD34+" = "Other B",
    "Macrophage" = "Myeloid"    
  )
)
```


```{r}
for (obj_name in names(objs)){
  for (category in names(label_mappings)) {
  objs[[obj_name]][[paste0('renamed.', category)]] <- 
    plyr::mapvalues(objs[[obj_name]][[category]] |> unlist() |> unname(),
                    from = names(label_mappings[[category]]),
                    to = label_mappings[[category]])
  }
}

```

# Marker genes

```{r}
celltype_markers <- read.csv('/fh/fast/_IRC/FHIL/grp/FHIL_knowledgebase/biology/celltype_markers.csv')

parse_marker_table <- function(celltype_dataframe) {
  celltype_dataframe %>%
    filter(expression_level == 'Increased' & tissue == 'PBMC') %>%
    filter(confidence != 'low') %>%
    filter(celltype %in% c('T', 'CD4+ T', 'CD8+ T', 'B', 'Monocyte', 'NK', 'Dendritic', 'pDC', 'Neutrophil', 'Non-classical monocyte', 'Classical monocyte')) %>%
    dplyr::select(gene_symbol, celltype) %>%
    unstack()
}

prune_marker_list <- function(input_list) {
  all_items <- unlist(input_list, use.names = TRUE)
  dup_items <- names(table(all_items))[table(all_items) > 1]
  input_list_cleaned <- lapply(input_list, function(x) setdiff(x, dup_items))
  
  new_groups <- sapply(dup_items, function(item) {
    found_in <- names(input_list)[sapply(input_list, function(x) item %in% x)]
    new_group_name <- paste(found_in, collapse = " &\n")
    return(new_group_name)
  }, USE.NAMES = TRUE)
  
  new_list <- input_list_cleaned  
  for (i in seq_along(dup_items)) {
    new_list[[new_groups[i]]] <- c(new_list[[new_groups[i]]], dup_items[i])
  }
  
  wrap_name <- function(name) {
    wrapped_name <- strwrap(name, width = 10, simplify = TRUE)
    wrapped_name <- paste(wrapped_name, collapse = "\n")
    return(wrapped_name)
  }
  
  names(new_list) <- sapply(names(new_list), wrap_name)# {
  new_list
}

celltype_markers <- parse_marker_table(celltype_markers)
celltype_markers <- prune_marker_list(celltype_markers)
```

```{r, fig.width=40, fig.height=12}
figures$marker_dotplots_pbmcsca <- lapply(objs, function(x) {
  DotPlot(x, celltype_markers, group.by = 'renamed.predicted.pbmcsca_seurat_annotations') + 
    theme_grey() +
    ggforestplot::geom_stripes()
})
figures$marker_dotplots_pbmcsca$Flex_F1A
```

```{r, fig.width=40, fig.height=12}
figures$marker_dotplots_pbmc3k <- lapply(objs, function(x) {
  DotPlot(x, celltype_markers, group.by = 'renamed.predicted.pbmc3k_seurat_annotations') + 
    theme_grey() +
    ggforestplot::geom_stripes()
})
figures$marker_dotplots_pbmc3k$Flex_F1A
```

```{r, fig.width=40, fig.height=12}
figures$marker_dotplots_mona <- lapply(objs, function(x) {
  DotPlot(x, celltype_markers, group.by = 'renamed.Mona.main.labels') + 
    theme_grey() +
    ggforestplot::geom_stripes()
})
figures$marker_dotplots_mona$Flex_F1A
```

```{r, fig.width=40, fig.height=12}
figures$marker_dotplots_nova <- lapply(objs, function(x) {
  DotPlot(x, celltype_markers, group.by = 'renamed.Novershtern.main.labels') + 
    theme_grey() +
    ggforestplot::geom_stripes()
})
figures$marker_dotplots_nova$Flex_F1A
```

```{r, fig.width=40, fig.height=12}
figures$marker_dotplots_hpca <- lapply(objs, function(x) {
  DotPlot(x, celltype_markers, group.by = 'renamed.HPCA.main.labels') + 
    theme_grey() +
    ggforestplot::geom_stripes()
})
figures$marker_dotplots_hpca$NextGEM3P_F1A
```

# Annotation cluster overlap

```{r, eval=FALSE}
tmp <- function(obj, ref1, ref2) {
  tabyl(obj@meta.data, .data[[ref1]], .data[[ref2]]) %>%
  # as.data.frame() %>%
  # reshape2::dcast(Var1 ~ Var2) %>%
  melt(, variable.name=ref2)
}
tmp1 <- tmp(obj, 'renamed.predicted.pbmc3k_seurat_annotations', 'renamed.Mona.main.labels') %>%
  mutate(value = value/sum(value))

ggplot(tmp1, aes(x=renamed.predicted.pbmc3k_seurat_annotations, y=renamed.Mona.main.labels, fill=value)) +
  geom_tile() +
  scale_fill_gradient(low='lightgrey', high='darkgreen', limits = c(0,1), oob=scales::squish) +
  theme(axis.text.x = element_text(angle=45, hjust=1))+
  labs(x='Ref 1', y='Ref 2', fill='log10 Cell count')
```

# Save objs

```{r}
saveRDS(objs, here('rds/5p/wt/04_objs_post_annotation.rds'), compress=FALSE)
```

# Render report

```{r, eval=TRUE}
## This only works if all the variables called in the format file 
## are in the current environment. Check the format file to tweak
## aesthetics 
rmarkdown::render(here('processing_scripts/5p/wt/04-reference_annotation.format.Rmd'),
                  output_file = '04-annotation.html',
                  output_dir = here('reports/5p/wt'))
```
