==> processing_scripts/5p/wt/05-clustering.format.Rmd <==
---
title: 'BM02 Clustering'
author: |
  | FHIL
  | Derrik Gratz
date: '`r Sys.Date()`'
output: 
  rmdformats::robobook:
    fig_width: 9
    toc_depth: 3
    css: '../../../config/robobook.css'
---

```{r, include=FALSE}
library(knitr)
knitr::opts_chunk$set(echo = FALSE,
                      fig.width = 8, fig.height = 6,
                      cache=FALSE, cache.lazy = FALSE) 
```

This report is intended as an in-depth look at the decision points, intermediate data, and possible points of interest. 
It does not represent a final state of the data, and not every figure here is intended to be used for publication. 
Aesthetics and image quality can be improved for final versions of images.

# Natural clustering {.tabset}

Clustering resolution for each sample was selected by optimizing for the silhouette score for a span of resolution values that
produced between 10 and 30 clusters.The Leiden algorithm implemented by iGraph was used for clustering.

Also looking at UMAP generation through 2 approaches: one with all genes, and one with
TCR, RB, and MT blacklisted from dimensional reduction. This may improve the overlap of samples,
but if it doesn't, we can just default to allowing all genes.

## Blacklisted {.tabset}

```{r, eval=TRUE, results='asis'}
out <- NULL
for (kit in unique(metadata$Kit)) {
    chunktext <- '\n### {{kit}} {.tabset .unlist}\n
    \n```{r, fig.width=6.5, fig.height=5}\n
    DimPlot(kit_level_objs[["{{kit}}"]], reduction="UMAP_blacklisted",
            group.by="orig.ident",
            raster = FALSE, raster.dpi = c(512, 512))
    \n```\n
    '
    out = c(out, knit_expand(text = chunktext))
}

cat(knit_child(text = out, quiet = TRUE), sep = '\n')
```

## All genes {.tabset}

```{r, eval=TRUE, results='asis'}
out <- NULL
for (kit in unique(metadata$Kit)) {
    chunktext <- '\n### {{kit}} {.tabset .unlist}\n
    \n```{r, fig.width=6.5, fig.height=5}\n
    DimPlot(kit_level_objs[["{{kit}}"]], reduction="UMAP_allgenes",
            group.by="orig.ident",
            raster = FALSE, raster.dpi = c(512, 512))
    \n```\n
    '
    out = c(out, knit_expand(text = chunktext))
}

cat(knit_child(text = out, quiet = TRUE), sep = '\n')
```

# Gene signatures {.tabset}

Celltype gene markers from literature. **If you have more, please let me know so we can expand the database**. 

```{r, eval=TRUE, results='asis'}
out <- NULL
for (kit in unique(metadata$Kit)) {
  # chunktext <- '\n## {{kit}} {.tabset .unlist}\n'
  # out = c(out, knit_expand(text = chunktext))
  # for (sample in metadata$Sample[metadata$Kit == kit]) {
  #   ind <- metadata$Individual[metadata$Sample==sample]
  #   rep <- metadata$Replicate[metadata$Sample==sample]
  chunktext <- '\n## {{kit}} {.tabset .unlist}\n
    \n```{r, fig.width=40, fig.height=11}\n
    figures$marker_dotplots[["{{kit}}"]]
    \n```\n
    '
    out = c(out, knit_expand(text = chunktext))
  # }
}

cat(knit_child(text = out, quiet = TRUE), sep = '\n')
```

# Reference mapping {.tabset}

Summarizing top annotations from all references used (see previous report for more details)

```{r, eval=TRUE, results='asis'}
out <- NULL
for (kit in unique(metadata$Kit)) {
  # chunktext <- '\n## {{kit}} {.tabset .unlist}\n'
  # out = c(out, knit_expand(text = chunktext))
  # for (sample in metadata$Sample[metadata$Kit == kit]) {
    # ind <- metadata$Individual[metadata$Sample==sample]
    # rep <- metadata$Replicate[metadata$Sample==sample]
    chunktext <- '\n## {{kit}} {.tabset .unlist}\n
      \n```{r, fig.width=12, fig.height=7}\n
      figures[["annotation_DT"]][["{{kit}}"]]
      \n```\n
      '
      out = c(out, knit_expand(text = chunktext))
  # }
}


cat(knit_child(text = out, quiet = TRUE), sep = '\n')
```

# Final annotations {.tabset}

My first pass at cell annotation, manually aggregating calls from the reference sets and marker genes.

**This may not be the final annotations**

```{r, eval=TRUE, results='asis'}
out <- NULL
for (kit in unique(metadata$Kit)) {
    chunktext <- '\n## {{kit}} {.unlist}\n
      \n```{r, fig.width=8, fig.height=5}\n
        DimPlot(kit_level_objs[["{{kit}}"]],
                group.by = "cell_labels.fine",
                raster = TRUE, raster.dpi = c(512, 512),
                label=TRUE)
        \n```\n
        \n```{r, fig.width=11, fig.height=5}\n
        DimPlot(kit_level_objs[["{{kit}}"]],
                group.by = "annotations",
                raster = TRUE, raster.dpi = c(512, 512),
                label=TRUE)
      \n```\n
      \n```{r}\n
      kit_level_objs[["{{kit}}"]]@meta.data$cell_labels.fine |> 
        tabyl() |> 
        arrange(desc(n)) |>
        adorn_rounding(digits=2) |>
        knitr::kable(col.names=c("Cell label", "Count", "Proportion"))
      \n```\n
      '
      out = c(out, knit_expand(text = chunktext))
}
cat(knit_child(text = out, quiet = TRUE), sep = '\n')
```
