==> processing_scripts/5p/wt/05-clustering.runfile.Rmd <==
---
title: 'SC processing'
author: "FHIL\nDerrikGratz"
date: '`r Sys.Date()`'
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      fig.width = 8, fig.height = 6,
                      cache=FALSE, cache.lazy = FALSE) 

library(tidyverse)  ## General R logic
library(here)       ## Easier specification of file locations
library(yaml)       ## parses config yaml
library(Seurat)
library(reshape2)   ## DF manipulation
# library(clustree)   ## Resolution selection for clustering
# library(ggraph)     ## To render clustree
library(bluster)     ## Sihouette scores
library(SingleR)    ## reference based annotation
library(ggforestplot) ## Striped lines for dotplots
library(SingleCellExperiment)
library(gridExtra)  ## arrange plots
library(janitor) ## Tabyl, adorn_rounding
library(ggrastr) ## Rasterize images for smaller size
library(DT)     ## Datatables
library(igraph) ## leiden clustering
rbindlist <- data.table::rbindlist
set.seed(33)
```

# Functions

```{r}
add_suffix_to_duplicates <- function(vec) {
  counts <- ave(seq_along(vec), vec, FUN = seq_along)
  vec[counts > 1] <- paste0(vec[counts > 1], "_", counts[counts > 1])
  return(vec)
}
```


# Load data

```{r}
objs <- readRDS(here('rds/5p/wt/04_objs_post_annotation.rds'))
metadata <- read.csv(here('config/5p/metadata.csv'))
figures <- list()
```

# UMAP

## Kit level

```{r}
blacklist_gene_patterns <- c(
'IG[HKL][VDJ]', 
'IGHM', 'IGHD', 'IGHE', 'IGHA[1-2]', 'IGHG[1-4]', 'IGKC', 'IGLC[1-7]', 'AC233755.1', #Ig constant genes 
'IGLL',
'TR[ABGD][CV]', #T-cell receptor genes
'^MT',
'^RP[SL]'
)

kit_level_objs <- lapply(setNames(unique(metadata$Kit), unique(metadata$Kit)), function(kit){
  samples <- metadata$Sample[metadata$Kit==kit]
  x <- merge(objs[[samples[1]]], objs[samples[2:length(samples)]], merge.data=TRUE)
  x[['RNA']] <- JoinLayers(x[['RNA']])
  x
})
kit_level_objs <- lapply(kit_level_objs, function(obj) {
  
  # Define blacklist 
  blacklist_genes <- c()
  for (pattern in blacklist_gene_patterns){
    blacklist_genes <- c(blacklist_genes, Features(obj)[grepl(pattern, Features(obj))])
  }
  blacklist_genes <- unique(blacklist_genes)
    
  obj <- obj |> 
    # NormalizeData() |> ## since merge.data is on
    # ScaleData() |>
    RunPCA(reduction.name='PCA_allgenes') |> 
    RunUMAP(reduction='PCA_allgenes', dims=1:20, reduction.name='UMAP_allgenes')
  
  whitelisted_genes <- 
    FindVariableFeatures(obj, selection.method = "vst",
                         nfeatures = 2000+length(blacklist_genes)) |>
    Seurat::VariableFeatures() |>
    setdiff(blacklist_genes)
  whitelisted_genes <- whitelisted_genes[1:2000]
  obj <- obj |>
    # ScaleData() |>
    RunPCA(features = whitelisted_genes, reduction.name='PCA_blacklisted') |> 
    RunUMAP(reduction='PCA_blacklisted',dims=1:20, reduction.name='UMAP_blacklisted')
  obj
})

# for (sample in unique(metadata$Sample)) {
#   kit <- metadata$Kit[metadata$Sample==sample]
#   umap <- kit_level_objs[[kit]]@reductions[['kit_level_UMAP']]
#   umap <- umap[Cells(objs[[sample]]),]
# }
# kit_level_umap_embeddings <- lapply(setNames(unique(metadata$Kit), unique(metadata$Kit)), function(kit){
#   
#   
# })
```

# Leiden Igraph

```{r igraph_funcs}
PrepareGraph <- function(g) {
  ## https://rdrr.io/github/joshpeters/westerlund/src/R/functions.R
  attributes(g)[[1]] <- NULL
  attributes(g)$class <- "dgCMatrix"
  g <- igraph::graph_from_adjacency_matrix(adjmatrix = g, mode = "undirected", weighted = TRUE, add.colnames = TRUE)
  g
}

relevel_clusters <- function(numbers) {
  value_counts <- table(numbers)
  sorted_values <- names(sort(value_counts, decreasing = TRUE))
  factor_map <- setNames(seq_along(sorted_values) - 1, sorted_values)
  replaced_values <- factor_map[as.character(numbers)]
  factor(replaced_values, levels = factor_map)
}

GroupSingletons <- function(ids, snn, min.size = 9, clusters.to.merge, group.singletons = TRUE, verbose = TRUE) {

  # usethis::ui_info("Merging small or provided clusters")

  # identify singletons
  singletons <- c()
  singletons <- names(x = which(x = table(ids) <= min.size))
  singletons <- intersect(x = unique(x = ids), singletons)

  if (!missing(clusters.to.merge)) {
    singletons <- c(singletons, as.character(clusters.to.merge))
  }

  if (!group.singletons) {
    ids[which(ids %in% singletons)] <- "singleton"
    return(ids)
  }

  # calculate connectivity of singletons to other clusters, add singleton
  # to cluster it is most connected to

  if (!is_empty(singletons)) {
    cluster_names <- as.character(x = unique(x = ids))
    cluster_names <- setdiff(x = cluster_names, y = singletons)
    connectivity <- vector(mode = "numeric", length = length(x = cluster_names))
    names(x = connectivity) <- cluster_names
    new.ids <- ids
    for (i in singletons) {
      i.cells <- names(which(ids == i))
      for (j in cluster_names) {
        j.cells <- names(which(ids == j))
        subSNN <- snn[i.cells, j.cells]
        # set.seed(1) # to match previous behavior, random seed being set in WhichCells
        if (is.object(x = subSNN)) {
          connectivity[j] <- sum(subSNN) / (nrow(x = subSNN) * ncol(x = subSNN))
        } else {
          connectivity[j] <- mean(x = subSNN)
        }
      }
      m <- max(connectivity, na.rm = T)
      mi <- which(x = connectivity == m, arr.ind = TRUE)
      closest_cluster <- sample(x = names(x = connectivity[mi]), 1)
      ids[i.cells] <- closest_cluster
    }
  }

  if (verbose) {
    message(paste(
      length(x = singletons),
      "singletons identified.",
      length(x = unique(x = ids)),
      "final clusters."
    ))
  }

  return(ids)
}

tune_leiden_resolution <- function(
  obj,          # A data frame or matrix of PC loadings
  ndim=20,
  graph_name='RNA_snn',
  reduction='pca',
  k=30,          # Target number of clusters
  start_resolution = 0.0001, # Initial resolution value
  initial_step = 0.0001,     # Initial step size for adjusting resolution
  max_iterations = 400,    # Maximum iterations to prevent infinite loops
  tolerance = 0         # Tolerance around target cluster number
) {
  current_resolution <- start_resolution
  step_size <- initial_step
  num_clusters <- 0
  iterations <- 0
  best_resolution <- current_resolution
  closest_difference <- Inf
  
  pca <- obj@reductions[[reduction]]@cell.embeddings[,1:ndim]
  graph <- PrepareGraph(obj@graphs[[graph_name]])
  
  while (iterations < max_iterations) {
    # Perform Leiden clustering with the current resolution
    clusters <- cluster_leiden(graph, resolution_parameter = current_resolution)
    ids <- clusters$membership
    names(ids) <- Cells(obj)
    ids <- GroupSingletons(ids, obj@graphs[[graph_name]])
    ids <- relevel_clusters(ids)
    names(ids) <- Cells(obj)
    clusters$membership <- ids
    num_clusters <- length(unique(clusters$membership))
    
    # Check if the current result is within the tolerance range
    difference <- abs(num_clusters - k)
    if (difference <= tolerance) {
      cat("Target number of clusters achieved within tolerance.\n")
      return(list(clusters = clusters, resolution = current_resolution))
    }
    
    # Update best resolution if the current attempt is closer to the target
    if (difference < closest_difference) {
      closest_difference <- difference
      best_resolution <- current_resolution
    }
    
    # Determine whether to increase or decrease resolution
    if (num_clusters < k) {
      if (num_clusters < k / 2) {
        step_size <- step_size * 1.5  # Increase step size to speed up convergence
      }
      current_resolution <- current_resolution + step_size
    } else {
      current_resolution <- current_resolution - step_size
      step_size <- step_size / 2  # Decrease step size to refine the search
    }
    
    # Increment iteration count
    iterations <- iterations + 1
  }
  # Return the best result if max iterations are reached
  cat("Max iterations reached. Returning the closest found result.\n")
  clusters <- cluster_leiden(graph, resolution_parameter = best_resolution)
  return(list(clusters = clusters, resolution = best_resolution))
}
```

```{r}
kit_level_objs <- lapply(kit_level_objs,
                         FindNeighbors, 
                         dims=1:20, reduction='PCA_allgenes',
                         compute.SNN=TRUE, #return.neighbor=TRUE, 
                         graph.name=c('RNA_nn', 'RNA_snn'))
```


```{r, eval=FALSE}
res_30k_kit_level <- lapply(kit_level_objs, tune_leiden_resolution, reduction='PCA_allgenes')
saveRDS(res_30k_kit_level, here('rds/5p/wt/leiden_sweep_res.rds'), compress=FALSE)
```

```{r}
res_30k_kit_level <- readRDS(here('rds/5p/wt/leiden_sweep_res.rds'))
```


```{r}
# for (smple in metadata$Sample) {
#   objs[[smple]] <- AddMetaData(objs[[smple]],
#                                metadata = res_30k[[smple]]$clusters$membership,
#                                col.name = 'clusters')
#   Idents(objs[[smple]]) <- 'clusters'
# }
for (kit in unique(metadata$Kit)) {
  kit_level_objs[[kit]] <- AddMetaData(kit_level_objs[[kit]],
                               metadata = res_30k_kit_level[[kit]]$clusters$membership,
                               col.name = 'clusters')
  Idents(kit_level_objs[[kit]]) <- 'clusters'
}
```

```{r}
lapply(kit_level_objs, DimPlot, label=TRUE, reduction='UMAP_blacklisted')
```


# Cluster annotations

## Reference label

```{r}
extract_cluster_annotations <- function(data, label_cols, score_cols, names) {
  results <- list()
  for (i in 1:length(label_cols)) {
    labels <- label_cols[i]
    scores <- score_cols[i]
    results[[names[i]]] <- data@meta.data %>%
      select(clusters, .data[[labels]], .data[[scores]]) %>%
      group_by(clusters, .data[[labels]]) %>%
      summarise(n = n(),
                med.score = round(median(.data[[scores]]),2), .groups = 'drop_last') %>%
      mutate(total_cells = sum(n),
             percent = round(n / total_cells * 100, 2)) %>%
      slice_max(n, with_ties = FALSE) %>%
      select(-n) %>%
    rename_with(~ paste0(names[i], '.', .x), .cols = c('percent', 'med.score'))
  }
  if (length(label_cols) > 1) {
    results <- Reduce(function(x,y) merge(x,y, by=c('clusters', 'total_cells')), results)
  } else {
    results <- results[[1]]
  }
  results <- results %>% arrange(clusters)
}
```

```{r}
cluster_annotations <- lapply(kit_level_objs, extract_cluster_annotations, 
                              label_cols = c("renamed.predicted.pbmcsca_seurat_annotations",
                                             'renamed.predicted.pbmc3k_seurat_annotations',
                                             "renamed.Mona.main.labels",
                                             "renamed.HPCA.main.labels"),
                              score_cols = c('predicted.pbmcsca_seurat_annotations.score',
                                             'predicted.pbmc3k_seurat_annotations.score',
                                             "Mona.main.delta.next",
                                             "HPCA.main.delta.next"),
                              names=c('seurat_pbmcsca', 'seurat_pbmc3k', 'singler_mona', 'singler_HPCA'))
```


```{r}
annotation_summary_DT <- function(data) {
  datatable(data, rownames = FALSE,
    options = list(pageLength = 22,
                   searching=FALSE,
                   ordering=FALSE,
                   lengthChange = FALSE), 
    colnames = c('Cluster', 'Cell count', 
                 'Seurat pbcmsca label', 'Seurat pbmcsca score', 'Seurat pbmcsca %', 
                 'Seurat pbcm3k label', 'Seurat pbmc3k score', 'Seurat pbmc3k %', 
                 'SingleR Mona label', 'SingleR Mona score', 'SingleR Mona %',
                 'SingleR HPCA label', 'SingleR HPCA score', 'SingleR HPCA %')
    ) %>%
  formatStyle(c("seurat_pbmcsca.percent", 'seurat_pbmc3k.percent',"singler_mona.percent","singler_HPCA.percent"),
              background = styleColorBar(range(c(0,100)), 'lightgreen'))  %>%
  formatStyle(c("seurat_pbmcsca.med.score",
                "seurat_pbmc3k.med.score",
                "singler_mona.med.score",
                "singler_HPCA.med.score"),
              background = styleColorBar(seq(0,0.9,.1), 'lightblue')) %>%
  formatStyle(
    c("renamed.predicted.pbmcsca_seurat_annotations",
      "renamed.predicted.pbmc3k_seurat_annotations",
    "renamed.Mona.main.labels",
    "renamed.HPCA.main.labels"),
    `border-left` = styleEqual(1, 'solid 3px')
  )
}
```

```{r}
kits <- unique(metadata$Kit)
names(kits) <- kits
figures[['annotation_DT']] <- lapply(kits, function(x) annotation_summary_DT(cluster_annotations[[x]]))
```

## Sample origin

```{r}
sample_origin_table <- lapply(kit_level_objs, function(x) {
  x@meta.data |>
    tabyl(clusters, orig.ident) |>
    adorn_percentages()
})
```

# Marker genes

```{r}
celltype_markers <- read.csv('/fh/fast/_IRC/FHIL/grp/FHIL_knowledgebase/biology/celltype_markers.csv')

parse_marker_table <- function(celltype_dataframe) {
  celltype_dataframe %>%
    filter(expression_level == 'Increased' & tissue == 'PBMC') %>%
    filter(confidence == 'high' | 
             (celltype %in% c('Dendritic','Monocyte', 
                              'Erythrocyte', 'Granulocyte',
                              'B naive', 'B memory') 
              & confidence == 'med')) %>%
    filter(celltype %in% c('T', 'CD4+ T', 'CD8+ T', 
                           'B', 'B naive', 'B memory',
                           'Monocyte', 'Non-classical monocyte', 'Classical monocyte',
                           'NK', 'Dendritic', 'pDC', 'Neutrophil',  
                           'Erythrocyte', 'Granulocyte',
                           'Megakaryocyte', 'Lymphocyte progenitor', 'HSPC'
                           )) %>%
    select(gene_symbol, celltype) %>%
    unstack()
}

prune_marker_list <- function(input_list) {
  all_items <- unlist(input_list, use.names = TRUE)
  dup_items <- names(table(all_items))[table(all_items) > 1]
  input_list_cleaned <- lapply(input_list, function(x) setdiff(x, dup_items))
  
  new_groups <- sapply(dup_items, function(item) {
    found_in <- names(input_list)[sapply(input_list, function(x) item %in% x)]
    new_group_name <- paste(found_in, collapse = " &\n")
    return(new_group_name)
  }, USE.NAMES = TRUE)
  
  new_list <- input_list_cleaned  
  for (i in seq_along(dup_items)) {
    new_list[[new_groups[i]]] <- c(new_list[[new_groups[i]]], dup_items[i])
  }
  
  wrap_name <- function(name) {
    wrapped_name <- strwrap(name, width = 13, simplify = TRUE)
    wrapped_name <- paste(wrapped_name, collapse = "\n")
    return(wrapped_name)
  }
  
  names(new_list) <- sapply(names(new_list), wrap_name)# {
  new_list
}

celltype_markers <- parse_marker_table(celltype_markers)
celltype_markers <- prune_marker_list(celltype_markers)
```

```{r}
# genes_to_test <- c(
#   'CLC', 'CPA3', 'GATA2', 'HDC', 'HPGDS', 'IL5RA', 'MS4A3'
# )
# DotPlot(kit_level_objs$Flex, features = genes_to_test, group.by='clusters')
# DotPlot(kit_level_objs$Scale, features = genes_to_test, group.by='clusters')
# DotPlot(kit_level_objs$Fluent_v4, features = genes_to_test, group.by='clusters')
```


```{r, fig.width=40, fig.height=12}
figures$marker_dotplots <- lapply(kit_level_objs, function(x) {
  z <- DotPlot(x, celltype_markers, group.by = 'clusters') + 
    theme_grey() +
    geom_stripes()
  z$data$feature.groups <- factor(z$data$feature.groups,
                                  levels=gtools::mixedsort(as.character(unique(z$data$feature.groups))))
  # z$data$feature.groups <- factor(z$data$feature.groups, levels =c(
  # 'B', 'B naive', 'B memory', 'Megakaryocyte',
  # 'T', "CD4+ T &\nT", 'CD4+ T', 'CD8+ T',
  # "CD8+ T &\nNK", 'NK', "CD8+ T &\nNK & pDC",
  # 'pDC', 'Dendritic', "Non-classical\nmonocyte", 
  # "Classical\nmonocyte", "Neutrophil", 'Erythrocyte',
  # 'Lymphocyte\nprogenitor', 'HSPC'))
  z
})
```

# Final labels

## NextGEM

```{r, fig.width=12}
kit <- 'NextGEM5P'
cell_labels <- c(
      'CD8+ T', #0
      'CD4+ T',
      'CD4+ T', 
      'CD4+ T',
      'CD4+ T',
      'CD4+ T',#5
      'CD4+ T',
      'CD4+ T',
      'T',
      'CD8+ T',
      'CD4+ T', #10
      'CD4+ T',
      'CD4+ T',
      'CD8+ T',
      'CD8+ T',
      'CD8+ T', #15
      'CD8+ T',
      'CD8+ T',
      'CD4+ T',
      'CD8+ T',
      'CD4+ T', #20
      'CD4+ T',
      'CD4+ T',
      'CD8+ T',
      'CD4+ T',
      'CD4+ T', #25
      'Monocyte',
      'B',
      'CD8+ T',
      'CD8+ T'
    )
kit_level_objs[[kit]]$cell_labels.fine <- 
  plyr::mapvalues(
    x = kit_level_objs[[kit]]$clusters,
    from = levels(kit_level_objs[[kit]]$clusters),
    to = cell_labels
  )

kit_level_objs[[kit]]$annotations <- 
  plyr::mapvalues(
    x = kit_level_objs[[kit]]$clusters,
    from = levels(kit_level_objs[[kit]]$clusters),
    to = add_suffix_to_duplicates(cell_labels)
  )

DimPlot(kit_level_objs[[kit]], group.by = 'annotations', label=TRUE)
DimPlot(kit_level_objs[[kit]], group.by = 'cell_labels.fine', label=TRUE)
```


## GEMX

```{r, fig.width=12, fig.height=5}
kit <- 'GEMX5P'
cell_labels <- c(
      'CD4+ T', #0
      'CD8+ T',
      'CD8+ T', 
      'CD8+ T',
      'CD4+ T',
      'CD4+ T',#5
      'CD4+ T',
      'CD8+ T',
      'T',
      'CD4+ T',
      'CD4+ T', #10
      'CD4+ T',
      'CD4+ T',
      'CD8+ T',
      'CD4+ T',
      'CD4+ T', #15
      'CD4+ T',
      'CD4+ T',
      'T',
      'CD4+ T',
      'CD4+ T', #20
      'CD8+ T',
      'CD8+ T',
      'CD8+ T',
      'T',
      'CD8+ T', #25
      'CD8+ T',
      'Monocyte',
      'B',
      'Megakaryocyte'
    )
kit_level_objs[[kit]]$cell_labels.fine <- 
  plyr::mapvalues(
    x = kit_level_objs[[kit]]$clusters,
    from = levels(kit_level_objs[[kit]]$clusters),
    to = cell_labels
  )

kit_level_objs[[kit]]$annotations <- 
  plyr::mapvalues(
    x = kit_level_objs[[kit]]$clusters,
    from = levels(kit_level_objs[[kit]]$clusters),
    to = add_suffix_to_duplicates(cell_labels)
  )

DimPlot(kit_level_objs[[kit]], group.by = 'annotations', label=TRUE)
DimPlot(kit_level_objs[[kit]], group.by = 'cell_labels.fine', label=TRUE)
```

## Parse

```{r, fig.width=10, fig.height=6}
kit <- 'Parse_v2'
cell_labels <- c(
      'CD4+ T', #0
      'CD8+ T',
      'CD8+ T',
      'CD8+ T',
      'CD4+ T',
      'CD4+ T', #5
      'CD4+ T',
      'CD4+ T',
      'CD8+ T',
      'CD8+ T',
      'CD4+ T', #10
      'CD4+ T',
      'CD4+ T',
      'CD4+ T',
      'CD4+ T',
      'CD8+ T', #15
      'CD4+ T',
      'CD4+ T',
      'CD4+ T',
      'CD4+ T',
      'CD8+ T', #20
      'CD8+ T',
      'CD4+ T',
      'CD8+ T',
      'CD4+ T',
      'CD8+ T', #25
      'Monocyte',
      'B',
      'CD4+ T',
      'T'
    )
kit_level_objs[[kit]]$cell_labels.fine <- 
  plyr::mapvalues(
    x = kit_level_objs[[kit]]$clusters,
    from = levels(kit_level_objs[[kit]]$clusters),
    to = cell_labels
  )

kit_level_objs[[kit]]$annotations <- 
  plyr::mapvalues(
    x = kit_level_objs[[kit]]$clusters,
    from = levels(kit_level_objs[[kit]]$clusters),
    to = add_suffix_to_duplicates(cell_labels)
  )

DimPlot(kit_level_objs[[kit]], group.by = 'annotations', label=TRUE)
DimPlot(kit_level_objs[[kit]], group.by = 'cell_labels.fine', label=TRUE)
```

# Save objs

```{r, eval=TRUE}
saveRDS(kit_level_objs, here('rds/5p/wt/05_merged_objs_post_clustering.rds'), compress=FALSE)
```

# Render report

```{r, eval=TRUE}
## This only works if all the variables called in the format file 
## are in the current environment. Check the format file to tweak
## aesthetics 
rmarkdown::render(here('processing_scripts/5p/wt/05-clustering.format.Rmd'),
                  output_file = '05-clustering.html',
                  output_dir = here('reports/5p/wt'))
```
